Extract events (#230)
* Extract events

* Use a macro to add From traits

* Expose TestEvent
