+++ packages/ciphernode/events/src/enclave_event/ciphernode_selected.rs
@@ -1,0 +1,21 @@
+use crate::E3id;
+use actix::Message;
+use serde::{Deserialize, Serialize};
+use std::fmt::{self, Display};
+
+#[derive(Message, Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[rtype(result = "()")]
+pub struct CiphernodeSelected {
+    pub e3_id: E3id,
+    pub threshold_m: usize,
+}
+
+impl Display for CiphernodeSelected {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "e3_id: {}, threshold_m: {}",
+            self.e3_id, self.threshold_m,
+        )
+    }
+}
