+++ packages/ciphernode/events/src/enclave_event/plaintext_aggregated.rs
@@ -1,0 +1,22 @@
+use crate::E3id;
+use actix::Message;
+use serde::{Deserialize, Serialize};
+use std::fmt::{self, Display};
+
+#[derive(Message, Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[rtype(result = "()")]
+pub struct PlaintextAggregated {
+    pub e3_id: E3id,
+    pub decrypted_output: Vec<u8>,
+    pub src_chain_id: u64,
+}
+
+impl Display for PlaintextAggregated {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "e3_id: {}, src_chain_id: {}",
+            self.e3_id, self.src_chain_id
+        )
+    }
+}
